requirements.txt
argparse==1.2.1
py==1.4.20
pytest==2.5.2
python-dateutil==2.2
six==1.6.1
wsgiref==0.1.2
